% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Diccionario para los datos de la ENCFT}
\format{
\link{list}
}
\source{
\url{https://github.com/endomer/encftr/tree/master/Files/Cuestionarios/}
}
\usage{
dict
}
\description{
Contiene las etiquetas tanto para las variables como para los valores de estas.
}
\keyword{datasets}
