% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encft.R
\docType{data}
\name{encft}
\alias{encft}
\title{Datos para simular la base de datos de la ENCFT}
\format{
\link{data.frame}
}
\source{
NA
}
\usage{
encft
}
\description{
Contiene un pequeño data.frame con la estructura básica necesaria para simular
la ENCFT. Útil para los ejemplos y viñetas del paquete.
}
\keyword{datasets}
