% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educacion.R
\name{ftc_alfabetizacion}
\alias{ftc_alfabetizacion}
\alias{ftc_compute_alfabetizacion}
\title{Alfabetización de la población}
\usage{
ftc_alfabetizacion(tbl, min_edad = 0, max_edad = Inf)

ftc_compute_alfabetizacion(tbl, min_edad = 0, max_edad = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame.}

\item{min_edad}{\link{numeric}: edad mínima del grupo para el que se quiere calcular
la alfabetización}

\item{max_edad}{\link{numeric}: edad máxima del grupo para el que se quiere calcular
la alfabetización}
}
\value{
el data.frame del input \code{tbl} con la variable \code{alfabetizacion} agregada
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Calcula la alfabetización de la población con edades
comprendidas entre \code{min_edad} y \code{max_edad}.
}
\examples{
\dontrun{
encft <- ftc_alfabetizacion(encft)
}
}
