% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_useLabels.R
\name{ftc_asLabels}
\alias{ftc_asLabels}
\title{Utilizar etiquetas de datos de la variable}
\usage{
ftc_asLabels(var)
}
\arguments{
\item{var}{variable o vector de datos que contiene las etiquetas a ser empleadas}
}
\value{
vector de datos con los valores cambiados por etiquetas
}
\description{
Reemplaza los valores de la variable por las etiquetas de esta
}
\examples{
\dontrun{
  encft$region <- ftc_asLabels(encft$region)
}
}
