% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ftc_browse_dict}
\alias{ftc_browse_dict}
\title{Visualizar el diccionario de forma interactiva}
\usage{
ftc_browse_dict()
}
\value{
Un elemento html con la información del diccionario
}
\description{
Visualizar el diccionario de forma interactiva
}
\examples{
\dontrun{
  ftc_browse_dict()
}
}
