% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_alfabetizacion.R
\name{ftc_compute_alfabetizacion}
\alias{ftc_compute_alfabetizacion}
\title{Alfabetización de la población}
\usage{
ftc_compute_alfabetizacion(tbl, edad_min = 0, edad_max = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame.}

\item{edad_min}{\link{numeric}: edad mínima del grupo para el que se quiere calcular
la alfabetización}

\item{edad_max}{\link{numeric}: edad máxima del grupo para el que se quiere calcular
la alfabetización}
}
\value{
el data.frame del input \code{tbl} con la variable \code{alfabetizacion} agregada
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Calcula la alfabetización de la poblaciín con edades
comprendidas entre \code{min_edad} y \code{max_edad}.
}
\examples{
\dontrun{
encft <- ftc_compute_alfabetizacion(encft)
}
}
