% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_anos_educacion.R
\name{ftc_compute_anos_educacion}
\alias{ftc_compute_anos_educacion}
\title{Años de educación}
\usage{
ftc_compute_anos_educacion(tbl, breaks = NULL, labels = NULL)
}
\arguments{
\item{tbl}{una conexión a base de datos o data.frame con los datos}

\item{breaks}{Puntos de corte para la variable. Vea \code{See also} para más detalles.}

\item{labels}{\link{character}: Nombres para los grupos de la variable.  Vea \code{See also} para más detalles.}
}
\value{
\link{data.frame}: Los datos suministrados en el argumento \code{tbl} con
la variable \code{anos_educacion} adicionada.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
  encft <- ftc_compute_anos_educacion(encft)
}
}
\seealso{
Vea \code{Dmisc::\link[Dmisc:cut3]{cut3}}
}
