% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_asistencia_escolar.R
\name{ftc_compute_asistencia_escolar}
\alias{ftc_compute_asistencia_escolar}
\title{Asistencia escolar}
\usage{
ftc_compute_asistencia_escolar(
  tbl,
  min_edad = 6,
  max_edad = 17,
  summer_fix = FALSE
)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos.}

\item{min_edad}{\link{numeric}: edad mínima del grupo para el que se desea calcular
la asistencia escolar}

\item{max_edad}{\link{numeric}: edad máxima del grupo para el que se desea calcular
la asistencia escolar}

\item{summer_fix}{\link{logical}: Indica si se ajusta por las personas que están
esperando el inicio de un nuevo año escolar.}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{asistencia_escolar} adicionada.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Calcula la asistencia escolar de los miembros del hogar entre \code{min_edad}
y \code{max_edad}.
}
\examples{
  encft <- data.frame(EDAD = c(10, 12, 9), TANDA_ASISTE = c(2, 9, 7),
                      MES = c(1, 4, 7), PORQUE_NO_ESTUDIA = c(3, 7, 1))
 ftc_compute_asistencia_escolar(encft)
 ftc_compute_asistencia_escolar(encft, summer_fix = TRUE)
}
