% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_factor_exp_anual.R
\name{ftc_compute_factor_exp_anual}
\alias{ftc_compute_factor_exp_anual}
\title{Factor de expansión anual}
\usage{
ftc_compute_factor_exp_anual(tbl)
}
\arguments{
\item{tbl}{Conexión a base de datos o dataframe con los datos.}
}
\value{
Conexión a base de datos o dataframe según input con la variable
\code{factor_exp_anual} agregada.
}
\description{
Factor de expansión anual
}
\examples{
  encft <- data.frame(FACTOR_EXPANSION = c(1028, 848, 400))
  ftc_compute_factor_exp_anual(encft)
}
