% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_grupos_edad.R
\name{ftc_compute_grupos_edad}
\alias{ftc_compute_grupos_edad}
\title{Grupos de edad}
\usage{
ftc_compute_grupos_edad(tbl, breaks = 20, labels = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos}

\item{breaks}{Puntos de corte para la variable. Vea \code{See also} para más detalles.}

\item{labels}{\link{character}: Nombres para los grupos de la variable.  Vea \code{See also} para más detalles.}
}
\value{
\link{data.frame}: Los datos suministrados en el argumento \code{tbl} con
la variable \code{grupos_edad} adicionada.
}
\description{
Divide la variable \code{EDAD} en el número o los grupos especificados.
}
\examples{
  encft <- data.frame(EDAD = c(0:100))
  ftc_compute_grupos_edad(encft)
}
\seealso{
Vea \code{Dmisc::\link[Dmisc:cut3]{cut3}}
}
