% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_hacinamiento.R
\name{ftc_compute_hacinamiento}
\alias{ftc_compute_hacinamiento}
\title{Nivel de hacinamiento del hogar}
\usage{
ftc_compute_hacinamiento(tbl, breaks = NULL, labels = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos}

\item{breaks}{Puntos de corte para la variable. Vea \code{See also} para más detalles.}

\item{labels}{\link{character}: Nombres para los grupos de la variable.  Vea \code{See also} para más detalles.}
}
\value{
\link{data.frame}: Los datos suministrados en el argumento \code{tbl} con
la variable \code{hacinamiento} adicionada.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Se define como el número de personas que habitan el hogar entre el número
de dormitorios de la vivienda. Se suele considerar hacinado un hogar con
2.5 o más personas por dormitorio.
}
\examples{
\dontrun{
encft <- ftc_compute_hacinamiento(encft)
}
}
\seealso{
Vea \code{Dmisc::\link[Dmisc:cut3]{cut3}}
}
