% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_pobreza_zona.R
\name{ftc_compute_pobreza_zona}
\alias{ftc_compute_pobreza_zona}
\title{Pobreza monetaria por zona de residencia}
\usage{
ftc_compute_pobreza_zona(tbl, conn, tbl_name = "pobreza_monetaria")
}
\arguments{
\item{tbl}{Un data.frame con los datos de la encuesta}

\item{conn}{Una conexi칩n a base de datos donde existe una tabla con los datos
de pobreza monetaria}

\item{tbl_name}{El nombre de la tabla en la base de datos;
type: character;
default: "pobreza_monetaria"}
}
\value{
los datos de la ENCFT m치s las variables de pobreza
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Esta funci칩n es temporal, entre tanto paso el script que genera pobreza
a una funci칩n R.
}
\examples{
\dontrun{
  encft <- ftc_compute_pobreza_zona(encft, conn)
}

}
