% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educacion.R
\name{ftc_compute_sobreedad_escolar}
\alias{ftc_compute_sobreedad_escolar}
\title{Sobreedad escolar}
\usage{
ftc_compute_sobreedad_escolar(tbl, nrezagos = 2)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos}

\item{nrezagos}{\link{numeric}: Número de años de rezago a considerar como sobreedad.
Vea \code{Seealso} para más detalles.}
}
\value{
\link{data.frame}: Los datos suministrados en el argumento \code{tbl} con
la variable \code{sobreedad_escolar} adicionada.
}
\description{
Alumnos que tienen más edad que la considerada teóricamente ideal, según
las normas legales de acceso al sistema educativo con respecto al año que cursa.
}
\examples{
\dontrun{
  encft <- ftc_compute_sobreedad_escolar(encft)
}
}
\seealso{
El Ministerio de Educación de la República Dominicana en el Anuario
de indicadores educativos 2017-2018 define la tasa de sobreedad como
Matriculados al inicio del año escolar en cada grado cuya edad es dos o más
años superiores a la edad legal o normal para el grado, como porcentaje del
total de matriculados en dicho grado.
}
