% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_tasa_alfabetizacion_hogar.R
\name{ftc_compute_tasa_alfabetizacion_hogar}
\alias{ftc_compute_tasa_alfabetizacion_hogar}
\title{Tasa de alfabetización del hogar}
\usage{
ftc_compute_tasa_alfabetizacion_hogar(tbl, edad_min = 0, edad_max = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame.}

\item{edad_min}{\link{numeric}: edad mínima del grupo para el que se quiere calcular
la tasa de alfabetización}

\item{edad_max}{\link{numeric}: edad máxima del grupo para el que se quiere calcular
la tasa de alfabetización}
}
\value{
el data.frame del input \code{tbl} con la variable \code{tasa_alfabetizacion_hogar} agregada
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Calcula la tasa de alfabetización del hogar para los miembros con edades
comprendidas entre \code{min_edad} y \code{max_edad}.
}
\examples{
\dontrun{
encft <- ftc_compute_tasa_alfabetizacion_hogar(encft)
}
}
