% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_tasa_dependencia.R
\name{ftc_compute_tasa_dependencia}
\alias{ftc_compute_tasa_dependencia}
\title{Tasa de dependencia del hogar}
\usage{
ftc_compute_tasa_dependencia(
  tbl,
  min_edad = 15,
  max_edad = 64,
  limit = c("both", "above", "below"),
  percent = TRUE,
  name_subfix = NULL,
  breaks = NULL,
  labels = NULL
)
}
\arguments{
\item{tbl}{\link{data.frame}: una dataframe con los datos de la ENCFT}

\item{min_edad}{\link{numeric}: edad mínima para el cálculo de dependencia}

\item{max_edad}{\link{numeric}: edad máxima para el cálculo de la dependencia}

\item{breaks}{si se especifica convierte la variable númerica resultante en
categorica. Vea \code{Dmisc::\link[Dmisc:cut3]{cut3}}}
}
\value{
un data frame con la variable \code{tasa_dependencia} agregada
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Calcula la tasa de dependencia de un hogar. Esto es, si se toma la edad mínima
de 15 años y máxima de 65 años, entonces calcula las personas con menos de
15 años o más de 65 años en el hogar como porcentaje de las personas con
entre 15 y 65 años del hogar.
}
\examples{
\dontrun{
encft <- ftc_compute_tasa_dependencia(encft)
}
}
