% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_trabajo_infantil.R
\name{ftc_compute_trabajo_infantil}
\alias{ftc_compute_trabajo_infantil}
\title{Niños de 5 a 14 años según condición laboral y escolar}
\usage{
ftc_compute_trabajo_infantil(tbl, summer_fix = FALSE)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos.}

\item{summer_fix}{\link{logical}: Indica si se ajusta por las personas que están
esperando el inicio de un nuevo año escolar.}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{trabajo_infantil} adicionada.
}
\description{
Calcula la cantidad de niños de 5 a 14 años según si estudian, trabajan,
ambas o ninguna.
}
\examples{
\dontrun{
  encft <- ftc_compute_trabajo_infantil(emcft)
}
}
