% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ftc_db_connect}
\alias{ftc_db_connect}
\alias{ftc_dbConnect}
\title{Conexión a base de datos}
\usage{
ftc_db_connect()

ftc_dbConnect()
}
\value{
Conexión a base de datos.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Vea \code{Dmisc::\link[Dmisc:db_connect]{db_connect}}
}
\examples{
\dontrun{
  conn <- ftc_db_connect()
}
}
