% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_dbCon.R
\name{ftc_dbConn}
\alias{ftc_dbConn}
\title{Conexion a bases de datos}
\usage{
ftc_dbConn(
  k_service_id = "postgre",
  db_name = "encft",
  db_host = "localhost",
  db_port = 5432
)
}
\arguments{
\item{k_service_id}{keyring service id. Vea \code{keyring::\link[keyring:key_list]{key_list}}}

\item{db_name}{nombre de la base de datos}

\item{db_host}{direccion del host de la base de datos}

\item{db_port}{puerto de la base de datos}
}
\value{
conexion de base de datos
}
\description{
Conexion a bases de datos
}
\examples{
\dontrun{
  conn <- ftc_dbConn()
}
}
