% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ftc_fuerza_trabajo_potencial}
\alias{ftc_fuerza_trabajo_potencial}
\title{Calcula la condici칩n de Fuerza de Trabajo Potencial}
\usage{
ftc_fuerza_trabajo_potencial(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexi칩n a base de datos o data.frame con los datos.}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{fuerza_trabajo_potencial} adicionada.
}
\description{
Comprende las personas que buscan trabajo pero no est치n disponible para trabajar,
o aquellos que no buscan y est치n disponibles para trabajar.
}
\examples{
\dontrun{
 encft <- ftc_fuerza_trabajo_potencial(encft)
}
}
