% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ftc_grupo_rama_pib}
\alias{ftc_grupo_rama_pib}
\title{Ramas de actividad utilizadas por el Banco Central en la clasificación del PIB}
\usage{
ftc_grupo_rama_pib(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos.}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{grupo_rama_pib} adicionada.
}
\description{
Ramas de actividad utilizadas por el Banco Central en la clasificación del PIB
}
\examples{
\dontrun{
encft <- ftc_grupo_rama_pib(encft)
}
}
