% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ftc_horas_semana}
\alias{ftc_horas_semana}
\title{Horas trabajadas a la semana}
\usage{
ftc_horas_semana(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos.}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{horas_semana} adicionada.
}
\description{
Calcula el número de horas trabajadas a la semana de la población ocupada.
Replica las horas publicadas por el Banco Central. Útil para el cálculo
del salario por hora publicado por el Banco Central.
}
\examples{
\dontrun{
  encft <- ftc_horas_semana(encft)
}
}
