% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_pobreza_monetaria.R
\name{ftc_ing_autoconsumo_autosuministro}
\alias{ftc_ing_autoconsumo_autosuministro}
\title{Title}
\usage{
ftc_ing_autoconsumo_autosuministro(tbl)
}
\arguments{
\item{tbl}{}
}
\description{
Title
}
