% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_pobreza_monetaria.R
\name{ftc_ing_mensual_ocup_prin_cuenta_propia}
\alias{ftc_ing_mensual_ocup_prin_cuenta_propia}
\title{Ingreso monetario por ocupacion principal para patronos y cuenta propia (Pobreza Monetaria)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ftc_ing_mensual_ocup_prin_cuenta_propia(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} data.frame con los datos de la ENCFT}
}
\value{
\link{data.frame} los datos suministrados en el data.frame \code{tbl}
con la variable \code{ing_mensual_ocup_prin_cuenta_propia} adicionada.
}
\description{
Ingreso monetario por ocupacion principal para patronos y cuenta propia (Pobreza Monetaria)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
encft <- ftc_ing_mensual_ocup_prin_cuenta_propia(encft)
}
}
