% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_pobreza_monetaria.R
\name{ftc_ing_mensual_ocup_prin_independiente}
\alias{ftc_ing_mensual_ocup_prin_independiente}
\title{Title}
\usage{
ftc_ing_mensual_ocup_prin_independiente(tbl)
}
\arguments{
\item{tbl}{}
}
\value{

}
\description{
Title
}
