% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_pobreza_monetaria.R
\name{ftc_ing_no_monetario_laboral}
\alias{ftc_ing_no_monetario_laboral}
\title{Title}
\usage{
ftc_ing_no_monetario_laboral(
  tbl,
  deflactar = TRUE,
  .keep = FALSE,
  .reuse = FALSE
)
}
\arguments{
\item{.reuse}{}
}
\description{
Title
}
