% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_pobreza_monetaria.R
\name{ftc_ing_pc_pobreza_def}
\alias{ftc_ing_pc_pobreza_def}
\title{Ingreso per c치pita deflactado (Pobreza Monetaria)}
\usage{
ftc_ing_pc_pobreza_def(tbl, .keep = FALSE, .reuse = FALSE)
}
\arguments{
\item{tbl}{DataFrame con los datos de la encuesta}

\item{.keep}{Lista de variables intermedias a mantener en el output.
Vea detalles en la funci칩n ftc_pobreza_monetaria.}

\item{.reuse}{Lista de variables intermedias a reutilizar del input.
Vea detalles en la funci칩n ftc_pobreza_monetaria.}
}
\value{
Los datos suministrados en el input con la variable \code{ing_pc_pobreza_def} agregada.
}
\description{
Ingreso per c치pita deflactado (Pobreza Monetaria)
}
\examples{
\dontrun{
ftc_ing_pc_pobreza_def(encft)
}
}
