% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_pobreza_monetaria.R
\name{ftc_ing_total_pobreza}
\alias{ftc_ing_total_pobreza}
\title{Title}
\usage{
ftc_ing_total_pobreza(tbl, .keep = FALSE, .reuse = FALSE)
}
\arguments{
\item{.reuse}{}
}
\value{

}
\description{
Title
}
