% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_pobreza_monetaria.R
\name{ftc_ing_transferencias_sociales}
\alias{ftc_ing_transferencias_sociales}
\title{Title}
\usage{
ftc_ing_transferencias_sociales(tbl, deflactar = TRUE)
}
\arguments{
\item{deflactar}{}
}
\value{

}
\description{
Title
}
