% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_pobreza_monetaria.R
\name{ftc_ing_utilidades_empresariales}
\alias{ftc_ing_utilidades_empresariales}
\title{Title}
\usage{
ftc_ing_utilidades_empresariales(tbl)
}
\arguments{
\item{tbl}{}
}
\value{

}
\description{
Title
}
