% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educacion.R
\name{ftc_matriculacion_escolar}
\alias{ftc_matriculacion_escolar}
\alias{ftc_compute_matriculacion_escolar}
\title{Matriculación escolar}
\usage{
ftc_matriculacion_escolar(tbl, min_edad = 6, max_edad = 17, summer_fix = FALSE)

ftc_compute_matriculacion_escolar(
  tbl,
  min_edad = 6,
  max_edad = 17,
  summer_fix = FALSE
)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos.}

\item{min_edad}{\link{numeric}: edad mínima del grupo para el que se desea calcular
la matriculación escolar}

\item{max_edad}{\link{numeric}: edad máxima del grupo para el que se desea calcular
la matriculación escolar}

\item{summer_fix}{\link{logical}: Indica si se ajusta por las personas que están
esperando el inicio de un nuevo año escolar.}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{matriculacion_escolar} adicionada.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Calcula la matriculación escolar de los miembros del hogar entre \code{min_edad}
y \code{max_edad}.
}
\examples{
  encft <- data.frame(EDAD = c(10, 12, 9), NIVEL_SE_MATRICULO = c(2, 9, 9),
                      MES = c(1, 4, 7), PORQUE_NO_ESTUDIA = c(3, 7, 1))
 ftc_matriculacion_escolar(encft)
 ftc_matriculacion_escolar(encft, summer_fix = TRUE)
}
