% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_recode_material_pared_exterior.R
\name{ftc_recode_material_pared_exterior}
\alias{ftc_recode_material_pared_exterior}
\title{Recode material pared exterior}
\usage{
ftc_recode_material_pared_exterior(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos.}
}
\value{
\link{data.frame}: Los datos suministrados en el input \code{tbl} con la
variable \code{MATERIAL_PARED_EXTERIOR} recodificada en función de la
variable \code{MATERIAL_PARED_EXTERIOR_ESP}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
La variable \code{MATERIAL_PARED_EXTERIOR_ESP} contiene muchas respuestas
que indican materialidad de calidad en la construcción de las paredes
exteriores de la vivienda.
}
\examples{
\dontrun{
  encft <- ftc_recode_material_pared_exterior(encft)
}
}
