% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_Labels.R
\name{ftc_set_Dict}
\alias{ftc_set_Dict}
\alias{ftc_set_labels}
\alias{ftc_setLabels}
\title{Asigna etiquetas de datos a las variables especificadas}
\usage{
ftc_set_Dict(tbl, dict = encftr::dict, subset = NULL, ...)

ftc_set_labels(tbl, dict = encftr::dict, vars = NULL)

ftc_setLabels(tbl, vars = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{dict}{\link{Dict}: Diccionario con las etiquetas de las variables}

\item{subset}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}

\item{...}{argumentos adicionales de \code{labeler::set_Dict}}
}
\value{
Los datos introducidos en el argumento \code{tbl} pero con etiquetas de datos
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Etiquetas de datos \code{vignette("labeler", package = "labeler")}
}
