% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_Labels.R
\name{ftc_setLabels}
\alias{ftc_setLabels}
\title{Asigna etiquetas de datos a las variables especificadas}
\usage{
ftc_setLabels(tbl, vars = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}
}
\value{
Los datos introducidos en el argumento \code{tbl} pero con etiquetas de datos
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Esta función permite asignar etiquetas de datos a las variables de un data.frame
o equivalente, a partir del diccionario de la encuesta.
}
\examples{
\dontrun{
  encft <- data.frame(ZONA = c(1, 2))
  str(encft)
  str(ftc_setLabels(encft))
}
}
\seealso{
Etiquetas de datos \code{vignette("labels", package = "endomer")}
}
