% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_setLabels.R
\name{ftc_setLabels_nivel_ultimo_ano_aprobado}
\alias{ftc_setLabels_nivel_ultimo_ano_aprobado}
\title{Etiquetas de datos variable nivel educativo}
\usage{
ftc_setLabels_nivel_ultimo_ano_aprobado(tbl)
}
\arguments{
\item{tbl}{Conexión a base de datos o dataframe con los datos de la ENCFT.}
}
\value{
Conexión a base de datos o dataframe según input.
}
\description{
Asigna las etiquetas de datos a la variable NIVEL_ULTIMO_ANO_APROBADO Se
puede utilizar por separado o a través de la variable \code{\link{ftc_setLabels}}
}
\examples{
\dontrun{
  encft <- ftc_setLabels_pobreza_zona(encft)
}
}
