% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_setLabels.R
\name{ftc_setLabels_sexo}
\alias{ftc_setLabels_sexo}
\title{Etiquetas de datos variable SEXO}
\usage{
ftc_setLabels_sexo(tbl)
}
\arguments{
\item{tbl}{Conexión a base de datos o dataframe con los datos de la ENCFT.}
}
\value{
Conexión a base de datos o dataframe según input.
}
\description{
Asigna las etiquetas de datos a la variable SEXO. Se puede utilizar por
separado o a través de la variable \code{\link{ftc_setLabels}}
}
\examples{
\dontrun{
  encft <- ftc_setLabels_sexo(encft)
}
}
