% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_sexo_jefe.R
\name{ftc_sexo_jefe}
\alias{ftc_sexo_jefe}
\title{Calcula el sexo del jefe del hogar}
\usage{
ftc_sexo_jefe(tbl)
}
\arguments{
\item{tbl}{datos de la base de datos de la ENCFT}
}
\value{
lo mismo que el parámetro tbl pero con una nueva columna \code{sexo_jefe}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
  \dontrun{
    ftc_sexo_jefe(tbl)
 }
}
