% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educacion.R
\name{ftc_tasa_alfabetizacion_hogar}
\alias{ftc_tasa_alfabetizacion_hogar}
\alias{ftc_compute_tasa_alfabetizacion_hogar}
\title{Tasa de alfabetización del hogar}
\usage{
ftc_tasa_alfabetizacion_hogar(tbl, min_edad = 0, max_edad = Inf)

ftc_compute_tasa_alfabetizacion_hogar(tbl, min_edad = 0, max_edad = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame.}

\item{min_edad}{\link{numeric}: edad mínima del grupo para el que se quiere calcular
la tasa de alfabetización}

\item{max_edad}{\link{numeric}: edad máxima del grupo para el que se quiere calcular
la tasa de alfabetización}
}
\value{
el data.frame del input \code{tbl} con la variable \code{tasa_alfabetizacion_hogar} agregada
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Calcula la tasa de alfabetización del hogar para los miembros con edades
comprendidas entre \code{min_edad} y \code{max_edad}.
}
\examples{
\dontrun{
encft <- ftc_tasa_alfabetizacion_hogar(encft)
}
}
