% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ftc_tiempo_total_empleo_dias}
\alias{ftc_tiempo_total_empleo_dias}
\title{Calcula el tiempo en el trabajo en días}
\usage{
ftc_tiempo_total_empleo_dias(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos.}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{tiempo_total_empleo_dias} adicionada.
}
\description{
Calcula el tiempo en el trabajo en días
}
\examples{
\dontrun{
encft <- ftc_tiempo_total_empleo_dias(encft)
}
}
