% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_useLabels.R
\name{ftc_useLabels}
\alias{ftc_useLabels}
\title{Utilizar etiquetas de datos de todas las variables}
\usage{
ftc_useLabels(tbl, vars = NULL)
}
\arguments{
\item{tbl}{Conexión a base de datos o dataframe con los datos de la ENCFT}

\item{vars}{Si es especificado, solo estas variables serán asignadas como
etiquetas de datos}
}
\value{
Conexión a base de datos o dataframe según input
}
\description{
Utilizar etiquetas de datos de todas las variables
}
\examples{
\dontrun{
  encft <- ftc_useLabels(encft)
}
}
