% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_Labels.R
\name{ftc_with_Dict}
\alias{ftc_with_Dict}
\alias{ftc_use_labels}
\alias{ftc_useLabels}
\title{Utiliza las etiquetas de datos de una variable/dataset}
\usage{
ftc_with_Dict(tbl, dict = NULL, subset = NULL, ...)

ftc_use_labels(tbl, dict = encftr::dict, vars = NULL, ...)

ftc_useLabels(tbl, dict = encftr::dict, vars = NULL, ...)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{dict}{\link{Dict}: Diccionario con las etiquetas de las variables}

\item{subset}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}

\item{...}{argumentos adicionales de \code{labeler::with_Dict}}
}
\value{
El dataset modificado
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
