% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{tipo_cambio_monedas}
\alias{tipo_cambio_monedas}
\title{Tipo de cambio de las monedas extranjeras publicadas por el Banco Central}
\format{
\link{data.frame}
}
\usage{
tipo_cambio_monedas
}
\description{
Tipo de cambio de las monedas extranjeras publicadas por el Banco Central
}
\keyword{datasets}
