% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_compute_pobreza_zona.R
\name{ftc_compute_pobreza_zona}
\alias{ftc_compute_pobreza_zona}
\title{Pobreza monetaria por zona de residencia}
\usage{
ftc_compute_pobreza_zona(tbl, conn)
}
\arguments{
\item{tbl}{los datos de la ENCFT}

\item{conn}{una conexi칩n a base de datos donde existe una tabla que se
llama pobreza_monetaria}
}
\value{
los datos de la ENCFT m치s las variables de pobreza
}
\description{
Esta funci칩n es temporal, entre tanto paso el script que genera pobreza
a una funci칩n R.
}
\examples{
\dontrun{
  encft <- ftc_compute_pobreza_zona(encft, conn)
}

}
