% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_regiones_desarrollo.R
\name{ftc_compute_region}
\alias{ftc_compute_region}
\title{Regiones de desarrollo}
\usage{
ftc_compute_region(tbl)
}
\arguments{
\item{tbl}{Conexión a base de datos o dataframe}
}
\value{
Conexión a base de datos o dataframe según input con la variable
\code{region} agregada
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Crea la variable Regiones de desarrollo según el decreto 710-04.
}
\examples{
\dontrun{
  encft <- ftc_compute_region(encft)
}
}
