% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_dbConnect.R
\name{ftc_dbConnect}
\alias{ftc_dbConnect}
\title{Conexión a base de datos}
\usage{
ftc_dbConnect()
}
\value{
Conexión a base de datos.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Vea \code{Dmisc::\link[Dmisc:dbConnect]{dbConnect}}
}
\examples{
\dontrun{
  conn <- ftc_dbConnect()
}
}
