% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ftc_perceptores_ingresos}
\alias{ftc_perceptores_ingresos}
\title{Población ocupada perceptora de ingresos}
\usage{
ftc_perceptores_ingresos(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o data.frame con los datos.}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{perceptores_ingresos} adicionada.
}
\description{
Población ocupada perceptora de ingresos
}
\examples{
\dontrun{
  encft <- ftc_perceptores_ingresos(encft)
}
}
