% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_setLabels.R
\name{ftc_setLabels}
\alias{ftc_setLabels}
\title{Etiquetas de datos}
\usage{
ftc_setLabels(tbl, vars = NULL)
}
\arguments{
\item{tbl}{Conexión a base de datos o dataframe con los datos de la ENCFT.}

\item{vars}{Si especificado, solo se asignaran las etiquetas a esas variables.}
}
\value{
Conexión a base de datos o dataframe según input.
}
\description{
Asigna etiquetas de datos a todas las variables que aplica.
}
\examples{
\dontrun{
  encft <- ftc_setLabels(encft)
}
}
