% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc_Labels.R
\name{ftc_useLabels}
\alias{ftc_useLabels}
\title{Utiliza las etiquetas de datos de una variable}
\usage{
ftc_useLabels(tbl, vars = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}
}
\value{
Los datos suministrados en el argumento \code{tbl}, pero en lugar de
valores utilizando las etiquetas de datos correspondientes
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{
\dontrun{
  encft <- data.frame(ZONA = c(1, 2))
  encft
  ftc_useLabels(encft)
}
}
\seealso{
Etiquetas de datos \code{vignette("labels", package = "endomer")}
}
